Corrected the return type based on the checkstyle warning

+++ src/main/java/com/javamvn/examples/Calculator.java
@@ -8,7 +8,7 @@ public class Calculator {
         return a-b;
     }
 
-    public long multiplication(int a,int b){
+    public int multiplication(int a,int b){
         return a*b;
     }
 }
